+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 ## [0.10.4] - 2021-03-25
-### Added
-- Checks for FASTQ files earlier in `treesapp abundance`
-
 ### Fixed
 - Checkpoint determination in `treesapp abundance`
+- '--report append' and 'report update' was not working properly in `treesapp abundance`.
+  Fixed by deduplicating PQueries prior to appending. 
+
+### Changed
+- Checks whether all FASTQ file paths exist earlier in `treesapp abundance`
 
 ## [0.10.3] - 2021-03-23
 ### Fixed